+++ src/main/java/sk/upjs/ics/entities/User.java
@@ -20,7 +20,7 @@ public class User {
     private String email;
     private String firstName;
     private String lastName;
-    private Float creditBalance;
+    private Long creditBalance;
     private String phone;
     private LocalDate birthDate;
     private boolean active;
@@ -60,7 +60,7 @@ public class User {
             user.setFirstName(rs.getString(prefix + "first_name"));
             user.setLastName(rs.getString(prefix + "last_name"));
             user.setEmail(rs.getString(prefix + "email"));
-            user.setCreditBalance(rs.getFloat(prefix + "credit_balance"));
+            user.setCreditBalance(rs.getLong(prefix + "credit_balance"));
             user.setPhone(rs.getString(prefix + "phone"));
 //            user.setBirthDate(rs.getDate(prefix + "birth_date").toLocalDate()); // todo
             user.setActive(rs.getBoolean(prefix + "active"));
